+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### 2025-08-20
 
+- Improve BLS12 precompile [#4073](https://github.com/lambdaclass/ethrex/pull/4073)
+
 - Improve blobbasefee opcode [#4092](https://github.com/lambdaclass/ethrex/pull/4092)
 
 ### 2025-08-19